library service created

+++ src/main/java/com/cis/batch33/library/model/Member.java
@@ -6,8 +6,7 @@ import java.util.List;
 
 @Data
 public class Member {
-
-    private long memberId;
+    private Integer memberId;
     private String firstName;
     private String lastName;
     private String emailAddress;
@@ -16,12 +15,3 @@ public class Member {
     private AddressDTO address;
     private List<CheckoutDTO> checkouts;
 }
-// dependencies are correct - pom.xml
-// connection properties -
-// create entity class and repository
-// Autowire in service class and update service methods
-
-
-// GET , PUT DELETE POST
-// book and member
-
